cache: end-to-end tests for cache
test-cache.pl drives client queries against a server.
It validates LOOKUP and LOOKUP_ONLY behavior.
I don't really care about UPLOAD behavior for now.

The CLIs now support an env var to control the location of the config file.
If not provided, defaults to cache/.config.json
This makes it easy to test on localhost.