client: npm: persistent cache
Problem:
HTTP queries are slow.
We want to make as few as possible.
The existing memory-only cache does not persist across different
instantiations of the module, so we forget things.

Solution:
Introduce a persistent cache. This is now the default.

Details:
The cache lives in os.tmpdir().
The expected use case is on a developer's machine, where the
same projects will be queried over and over.
Any previously-answered queries will be resolved locally.
Only newly/recently-added regexes need contact the server.

Cached results do not expire.
If we add new detectors or otherwise change the definition of vulnerability
on the server side, we should add expiration into the client-side cache.

Fix:
This fixes #26.