detect: rxxr2 parser mis-handled empty suggestions
Problem:
When no suffix or no prefix is required, rxxr2 emits an empty string.
The output parsing regex was looking for '(.+)' which
rejects the empty string.

Solution:
Switch to '(.*)'.

Test:
I added a new test case.
However, it appears that run-tests.pl is not precise enough
to notice the parse failure.
This is unfortunate.